+++ src/openvpn/ssl.c
@@ -345,7 +345,6 @@ void
 free_ssl_lib(void)
 {
     crypto_uninit_lib();
-    prng_uninit();
 
     tls_free_lib();
 }